\name{butterfly}
\alias{butterfly}
\docType{data}
\title{Fisher's butterfly data}
\description{The famous Fisher's butterfly data originally appeared in Fisher 1943. It has been re-analyzed in
many publications in the literature.}
%\author{Ji-Ping Wang, Department of Statistics, Northwestern University Maintainer: \email{jzwang@northwestern.edu}}
\references{
Fisher, R. A., Corbet, A. S., andWilliams, C. B. ,1943, The Relation Between the Number of Species and the Number of Individuals in a Random Sample of an Animal Population, Journal of Animal Ecology, 12, 42-58.
}

\examples{
##load library
library(SPECIES)

##load data that coming with the package.
data(butterfly)
}
