\name{cottontail}
\alias{cottontail}
\docType{data}
\title{Cottontail data}
\description{The cottontail data was analyzed in Chao 1987}
%\author{Ji-Ping Wang, Department of Statistics, Northwestern University, Maintainer: \email{jzwang@northwestern.edu}}
\references{
Chao, A. (1987). Estimating the population size for capture-recapture data with unequal catchability. Biometrics
43, 783-791. 
}

\examples{
##load library
library(SPECIES)

##load data that coming with the package.
data(cottontail)
}
