\name{EST}
\alias{EST}
\docType{data}
\title{EST data}
\description{The Arabidopsis thaliana expressed sequence tag (EST) data  originally appeared in Wang and Lindsay 2005. It was recently reanalyzed in Wang 2010. For convenience, the frequency at \eqn{j=17}  is used to denote the total count of species with \eqn{j\ge17}.}
%\author{Ji-Ping Wang, Department of Statistics, Northwestern University, Maintainer: \email{jzwang@northwestern.edu}}
\references{
Wang, J.-P. Z. and Lindsay, B. G. ,(2005), A penalized nonparametric maximum likelilhood approach to species richness estimation. Journal of American Statistical Association, 2005,100(471):942-959
}

\examples{
##load library
library(SPECIES)

##load data that coming with the package.
data(EST)
}
