\name{insects}
\alias{insects}
\docType{data}
\title{Insects data}
\description{The insects data was analyzed in Burnham and Overton 1979. The frequency at \eqn{j=6} is used to denote the total count of species with \eqn{j\ge 6}.}
%\author{Ji-Ping Wang, Department of Statistics, Northwestern University, Maintainer: \email{jzwang@northwestern.edu}}
\references{
Burnham, K. P., and Overton,W. S. (1979), Robust Estimation of Population Size When Capture Probabilities Vary Among Animals, Ecology, 60, 927-936.
}

\examples{
##load library
library(SPECIES)

##load data that coming with the package.
data(insects)
}
