\name{microbial}
\alias{microbial}
\docType{data}
\title{Microbial species data}
\description{The microbial species data originally appeared in Acinas et al 2004. Recently
it was re-analyzed by Bohning and Schon 2005, and Wang 2009.}
%\author{Ji-Ping Wang, Department of Statistics, Northwestern University}
\references{
Acinas, S., Klepac-Ceraj, V., Hunt, D., Pharino, C., Ceraj, I., Distel, D., and Polz, M. (2004), Fine-scale phylogenetic architecture of a complex bacterial community. Nature, 430, 551-554.

Hong, S. H., and Bunge, J. and Jeon, S.O. and Epstein, S. (2006), Predicting microbial species richness, Proc. Natl. Acad. Sci, 103, 117-122.
}

\examples{
##load library
library(SPECIES)

##load data that coming with the package.
data(microbial)
}
